Use dashboard-slug if specified in module stub.
This is so that we don't lose functionality for existing dashboards
which do not fit nicely into the department/agency/service hierarchy.
Namely, the G-Cloud and site activity dashboards.

We may want to look at a better way to manage this in future, but
for the moment, use dashboard-slug if it is defined.